+++ pkg/usagestats/stats.go
@@ -71,8 +71,19 @@ func sendReport(ctx context.Context, seed *ClusterSeed, interval time.Time) erro
 	return nil
 }
 
-// buildReport builds the report to be sent to the stats server
+// buildReport builds the report to be sent to the stats server,
+// this report includes the cluster seed data.
 func buildReport(seed *ClusterSeed, interval time.Time) Report {
+	report := BuildStats()
+	report.ClusterID = seed.UID
+	report.CreatedAt = seed.CreatedAt
+	report.Interval = interval
+
+	return report
+}
+
+// BuildStats builds the report without cluster seed data
+func BuildStats() Report {
 	var (
 		targetName  string
 		editionName string
@@ -89,10 +100,7 @@ func buildReport(seed *ClusterSeed, interval time.Time) Report {
 	}
 
 	return Report{
-		ClusterID:         seed.UID,
 		PrometheusVersion: build.GetVersion(),
-		CreatedAt:         seed.CreatedAt,
-		Interval:          interval,
 		IntervalPeriod:    reportInterval.Seconds(),
 		Os:                runtime.GOOS,
 		Arch:              runtime.GOARCH,
@@ -167,7 +175,7 @@ func NewFloat(name string) *expvar.Float {
 }
 
 // NewInt returns a new Int stats object.
-// If an Int stats object object with the same name already exists it is returned.
+// If an Int stats object with the same name already exists it is returned.
 func NewInt(name string) *expvar.Int {
 	existing := expvar.Get(statsPrefix + name)
 	if existing != nil {